pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "faros"
version = "0.1.0"
description = "Faros Context Orchestrator: Unified DevFinOps Intelligence CLI"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "duckdb>=0.9.0",
    "dspy-ai>=2.0.0",
    "mcp>=0.1.1",
    "sentence-transformers>=2.2.2",
    "PyYAML>=6.0.1",
    "argparse"
]

[project.scripts]
faros = "faros.cli:main"
